+++ tests/test_models/test_mlp.py
@@ -24,6 +24,7 @@ def _integration(data_with_covariates, tmp_path, gpus, data_loader_kwargs={}, tr
     dataloaders_with_covariates = make_dataloaders(data_with_covariates, **data_loader_default_kwargs)
     train_dataloader = dataloaders_with_covariates["train"]
     val_dataloader = dataloaders_with_covariates["val"]
+    test_dataloader = dataloaders_with_covariates["test"]
     early_stop_callback = EarlyStopping(
         monitor="val_loss", min_delta=1e-4, patience=1, verbose=False, mode="min", strict=False
     )
@@ -39,6 +40,7 @@ def _integration(data_with_covariates, tmp_path, gpus, data_loader_kwargs={}, tr
         default_root_dir=tmp_path,
         limit_train_batches=2,
         limit_val_batches=2,
+        limit_test_batches=2,
         logger=logger,
     )
 
@@ -60,6 +62,9 @@ def _integration(data_with_covariates, tmp_path, gpus, data_loader_kwargs={}, tr
 
         # check prediction
         net.predict(val_dataloader, fast_dev_run=True, return_index=True, return_decoder_lengths=True)
+        # check test dataloader
+        test_outputs = trainer.test(net, test_dataloaders=test_dataloader)
+        assert len(test_outputs) > 0
     finally:
         shutil.rmtree(tmp_path, ignore_errors=True)
 
@@ -86,7 +91,6 @@ def _integration(data_with_covariates, tmp_path, gpus, data_loader_kwargs={}, tr
         ),
         dict(optimizer="SGD", weight_decay=1e-3),
         dict(optimizer=lambda params, lr: SGD(params, lr=lr, weight_decay=1e-3)),
-        dict(optimizer=SGD, weight_decay=1e-4),
     ],
 )
 def test_integration(data_with_covariates, tmp_path, gpus, kwargs):