+++ CHANGELOG.md
@@ -4,13 +4,18 @@
 
 ### Added
 
+- Added support for running `pytorch_lightning.trainer.test` (#759)
+
 ### Fixed
 
 - Fix inattention mutation to `x_cont` (#732).
+- Compatability with pytorch-lightning 1.5 (#758)
 
 ### Contributors
 
 - eavae
+- danielgafni
+- jdb78
 
 ## v0.9.1 Maintenance Release (26/09/2021)